requirements.txt
pytest
pytest-playwright
Faker